important, lets see if the tag works this time. a few things, sorry they're all in 1, but stupid git can't do line by line commits if the lines are too close apparently :( here goes:
first, moved the return false of the render call to the top, making it more readable, and less code again, brevity ftw :)

next, took out init function in controller, not needed for anything in core and the same thing can be achieved by extending a controller and overwriting the constructor. Seems to reduce code size without effecting functionality. changed call in application to reflect this.

next, did a stack trace for render calls:

#0 teth/core/templates/CoreTemplate.php(59): include()
#1 teth/core/controllers/CoreController.php(39): CoreTemplate->content()
#2 teth/core/templates/CoreTemplate.php(94): CoreController->execute()
#3 teth/core/applications/CoreApplication.php(82): CoreTemplate::render()

that's the general pattern. seems odd that we jump back and forth between controller and template, and with no good reason since all the controller execute seems to do for now is run the action and jump back to template to do output. I moved this logic into the render call itself, so it checks for action method existence and runs it if it's there.